+++ kernel/packages/shared/comms/index.ts
@@ -9,7 +9,8 @@ import {
 import { CommunicationsController } from 'shared/apis/CommunicationsController'
 import { defaultLogger } from 'shared/logger'
 import { ChatMessage as InternalChatMessage, ChatMessageType, SceneFeatureToggles } from 'shared/types'
-import { positionObservable, PositionReport, lastPlayerPosition, lastPlayerScene } from 'shared/world/positionThings'
+import { positionObservable, PositionReport, lastPlayerPosition } from 'shared/world/positionThings'
+import { lastPlayerScene } from 'shared/world/sceneState'
 import { ProfileAsPromise } from '../profiles/ProfileAsPromise'
 import { notifyStatusThroughChat } from './chat'
 import { CliBrokerConnection } from './CliBrokerConnection'
